parity_db
Function migrate

pub fn migrate(
    from: &Path,
    to: Options,
    overwrite: bool,
    force_migrate: &[u8],
) -> Result<()>

Attempt to migrate a database to a new configuration with different column settings. from Source database path to New database configuration. overwrite Ignore path set in to and attempt to overwrite data in place. This may be faster but if migration fails data may be lost force_migrate Force column re-population even if its setting did not change.

Note that migration between hash to btree columns is not possible.