Crate parity_db

Structs

BTreeIterator

CheckOptions

Options for producing a database dump.

ColumnOptions

Database column configuration.

ColumnStatSummary

Column statistics summary.

Db

Database instance.

Options

Database configuration.

StatSummary

Database statistics summary.

ValueIterState

Value iteration state

Enums

CompressionType

Different compression type allowed and their u8 representation.

Error

Operation

Different operations allowed for a commit. Behavior may differs depending on column configuration.

Constants

KEY_SIZE

Functions

clear_column

Clear specified column. All data is removed and stats are reset. Database must be closed before calling this.

migrate

Attempt to migrate a database to a new configuration with different column settings. from Source database path to New database configuration. overwrite Ignore path set in to and attempt to overwrite data in place. This may be faster but if migration fails data may be lost force_migrate Force column re-population even if its setting did not change.

Type Aliases

ColId

Key

Result

Value

Value is just a vector of bytes. Value sizes up to 4Gb are allowed.